# frontend/src/scheme/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/db-test": {
      "get": {
        "summary": "Test Db Connection",
        "operationId": "test_db_connection_api_db_test_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/practice/start": {
      "get": {
        "summary": "Start Practice",
        "operationId": "start_practice_api_practice_start_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/practice/settings": {
      "post": {
        "summary": "Save Settings",
        "operationId": "save_settings_api_practice_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practice/question/{practice_id}": {
      "get": {
        "summary": "Get Question",
        "operationId": "get_question_api_practice_question__practice_id__get",
        "parameters": [
          {
            "name": "practice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Practice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionResponse"
                  },
                  "title": "Response Get Question Api Practice Question  Practice Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practice/answer/{question_id}": {
      "post": {
        "summary": "Submit Answer",
        "operationId": "submit_answer_api_practice_answer__question_id__post",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerSubmit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practice/results/{practice_id}": {
      "get": {
        "summary": "Get Results",
        "operationId": "get_results_api_practice_results__practice_id__get",
        "parameters": [
          {
            "name": "practice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Practice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnswerResponse": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "score",
          "feedback"
        ],
        "title": "AnswerResponse"
      },
      "AnswerSubmit": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "AnswerSubmit"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PracticeResult": {
        "properties": {
          "total_score": {
            "type": "number",
            "title": "Total Score"
          },
          "individual_scores": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Individual Scores"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "total_score",
          "individual_scores",
          "feedback"
        ],
        "title": "PracticeResult"
      },
      "PracticeSettings": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "toefl_score": {
            "type": "string",
            "title": "Toefl Score"
          },
          "num_questions": {
            "type": "integer",
            "title": "Num Questions"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "toefl_score",
          "num_questions",
          "duration"
        ],
        "title": "PracticeSettings"
      },
      "PracticeSettingsResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "practice_id": {
            "type": "integer",
            "title": "Practice Id"
          }
        },
        "type": "object",
        "required": [
          "message",
          "practice_id"
        ],
        "title": "PracticeSettingsResponse"
      },
      "QuestionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content"
        ],
        "title": "QuestionResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}